Dedup the code and styles
