+++ frontend/.gitignore
@@ -5,9 +5,7 @@ node_modules/
 .vite/
 
 # Build output
-dist/*
-!dist/.gitkeep
-!dist/inventario.txt
+dist/
 *.local
 
 # Logs
