+++ frontend/src/assets/_components.scss
@@ -15,11 +15,17 @@
   display: inline-block;
   padding: 0.75rem 1.5rem;
   border-radius: $default-radius;
+  cursor: pointer;
   font-weight: 500;
   font-size: 1rem; /* Ensure consistent font size */
   text-align: center;
   transition: background-color 0.2s, color 0.2s;
   border: none;
+
+  &:disabled {
+    opacity: 0.65;
+    cursor: not-allowed;
+  }
 
   &-primary {
     background-color: $primary-color;
@@ -32,13 +38,23 @@
   }
 
   &-secondary {
-    background-color: $light-bg-color;
-    color: $text-color;
-    border: 1px solid $border-color;
+    background-color: $secondary-color;
+    color: white;
+    //border: 1px solid $border-color;
 
     &:hover {
-      background-color: $light-hover-bg-color;
+      background-color: $secondary-hover-color;
       text-decoration: none;
+    }
+
+    &-alt {
+      background-color: $light-bg-color;
+      color: $text-color;
+      border: 1px solid $border-color;
+
+      &:hover {
+        background-color: $light-hover-bg-color;
+      }
     }
   }
