+++ frontend/src/assets/_variables.scss
@@ -2,6 +2,7 @@
 $primary-color: #4CAF50;
 $primary-hover-color: #43a047;
 $secondary-color: #2c3e50;
+$secondary-hover-color: #1a242f; // color.adjust($secondary-color, $lightness: -10%);
 $light-bg-color: #f8f9fa;
 $light-hover-bg-color: #e9ecef;
 $danger-color: #dc3545;
