+++ frontend/src/assets/primevue-dropdown.scss
@@ -15,7 +15,7 @@
   --p-select-focus-ring-style: solid;
   --p-select-focus-ring-color: rgb(76 175 80 / 20%); /* Primary color with opacity */
   --p-select-focus-ring-offset: 2px;
-  --p-select-disabled-background: #f8f9fa; /* Match light-bg-color from main.scss */
+  --p-select-disabled-background: #f8f9fa; /* Match light-bg-colorx from main.scss */
   --p-select-clear-icon-color: #6c757d; /* Match text-secondary-color from main.scss */
   --p-select-dropdown-color: #6c757d; /* Match text-secondary-color from main.scss */
   --p-select-dropdown-width: 2.5rem;
